Extract some modules
