+++ nym-vpnd/src/cli.rs
@@ -1,0 +1,29 @@
+use clap::Parser;
+use nym_vpn_lib::nym_bin_common::bin_info_local_vergen;
+use std::{path::PathBuf, sync::OnceLock};
+
+// Helper for passing LONG_VERSION to clap
+fn pretty_build_info_static() -> &'static str {
+    static PRETTY_BUILD_INFORMATION: OnceLock<String> = OnceLock::new();
+    PRETTY_BUILD_INFORMATION.get_or_init(|| bin_info_local_vergen!().pretty_print())
+}
+
+#[derive(Parser)]
+#[clap(author = "Nymtech", version, about, long_version = pretty_build_info_static())]
+pub(crate) struct CliArgs {
+    /// Path pointing to an env file describing the network.
+    #[arg(short, long, value_parser = check_path)]
+    pub(crate) config_env_file: Option<PathBuf>,
+}
+
+fn check_path(path: &str) -> Result<PathBuf, String> {
+    let path = PathBuf::from(path);
+    if !path.exists() {
+        return Err(format!("Path {:?} does not exist", path));
+    }
+    if !path.is_file() {
+        return Err(format!("Path {:?} is not a file", path));
+    }
+    Ok(path)
+}
+
