+++ nym-vpnd/src/command_interface.rs
@@ -33,6 +33,26 @@ pub fn start_command_interface(
                     .await
             });
 
+            // Using TaskManager::catch_interrupt() here is a bit of a hack that we use for now.
+            // The real solution is to:
+            //
+            // 1. Register signal handler as a separate task. This handler will need the ability to
+            //    signal shutdown, which TaskClient currently doesn't have
+            //
+            // 2. Register error message handler that listens for errors / drops of the tasks, this
+            //    will also signal shutdown.
+            //
+            // 3. (TaskManager needs to be updated so that the vpn-lib does not wait for signals.
+            //    Or maybe we can just not call catch_interrupt() at all in vpn-lib.)
+            //
+            // 4. Then we need to wait for all tasks to finish, like in
+            //    TaskManager::wait_for_shutdown(). Since this also listens to ctrl-c, we can't
+            //    start this until shutdown has already been signalled.
+            //
+            // 5. Back up at the top-level where we join the threads, we check for return errors
+            //    and handle that there.
+
+            // This call will:
             // Wait for interrupt
             // Send shutdown signal to all tasks
             // Wait for all tasks to finish
